tests/wpt/web-platform-tests/FileAPI/blob/Blob-slice-overflow.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Blob slice overflow</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="help" href="https://w3c.github.io/FileAPI/#dfn-slice">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>

var text = '';

for (var i = 0; i < 2000; ++i) {
  text += 'A';
}

test(function() {
  var blob = new Blob([text]);
  var sliceBlob = blob.slice(-1, 2000);
  assert_equals(sliceBlob.size, 2000-(2000-1), "Bolb slice size");
}, "slice start is negative, relativeStart will be max((size + start), 0)");

test(function() {
  var blob = new Blob([text]);
  var sliceBlob = blob.slice(2001, 2000);
  assert_equals(sliceBlob.size, 0, "Bolb slice size");
}, "slice start is greater than blob size, relativeStart will be min(start, size)");

test(function() {
  var blob = new Blob([text]);
  var sliceBlob = blob.slice(1998, -1);
  assert_equals(sliceBlob.size, (2000-1)-1998, "Bolb slice size");
}, "slice end is negative, relativeEnd will be max((size + end), 0)");

test(function() {
  var blob = new Blob([text]);
  var sliceBlob = blob.slice(1998, 2999);
  assert_equals(sliceBlob.size, 2000-1998, "Bolb slice size");
}, "slice end is greater than blob size, relativeEnd will be min(end, size)");

</script>
